Delete file names and create new to avoid file issues
