+++ src/lambda/database/data/index.js
@@ -2,7 +2,7 @@ import mongoose from 'mongoose';
 import dbConnection from '../dbConnection';
 import resetDb from '../resetDb';
 import users from './users';
-import rentalData from './rentalData';
+import rentalRecords from './rentalRecords';
 
 require('dotenv').config();
 
@@ -11,7 +11,7 @@ const buildData = () =>
     try {
       await resetDb();
       await users();
-      await rentalData();
+      await rentalRecords();
     } catch (err) {
       // eslint-disable-next-line no-console
       console.log('err during building the test db, try again', err);
