+++ src/lambda/database/resetDb.js
@@ -1,10 +1,10 @@
-import User from './models/users';
-import RentalData from './models/RentalData';
+import User from './models/User';
+import RentalRecord from './models/RentalRecord';
 
 const resetDB = async () => {
   try {
     await User.deleteMany();
-    return RentalData.deleteMany();
+    return RentalRecord.deleteMany();
   } catch (err) {
     // eslint-disable-next-line no-console
     console.log('Error during resting the db, try again', err);
